messaging-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
kafka-python==2.0.2
grpcio==1.62.2
grpcio-tools==1.62.2
pydantic==2.7.1
python-multipart==0.0.9
websockets==12.
python-jose==3.3.0